+++ src/test/kotlin/lotto/LottoTest.kt
@@ -1,6 +1,6 @@
 package lotto
 
-import lotto.models.Lotto
+import lotto.Lotto
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.assertj.core.api.Assertions.assertThat
