+++ src/test/kotlin/lotto/models/WinningLottoTest.kt
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
+import lotto.Lotto
+
 class WinningLottoTest {
 
     @Test
